ADH-4718: renamed some classes
